Check if generator burns fluid #385
Previously, there was no way to tell if a generator burns fluid or uses the fluid's temperature. This has since been added to the base game.
More places where we need to check the generator rather than the fuel.

Py has a fluid (combustion mixture) which has a fuel value but that shouldn't be used. It's temperature should be used instead.